requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
Django==1.11.1
django-cors-headers==2.1.0
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.1
djangorestframework==3.6.3
html5lib==0.999999999
idna==2.5
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
six==1.10.0
Unidecode==0.4.20
urllib3==1.21.1
wagtail==1.10.1
wagtailfontawesome==1.0.6
webencodings==0.5.1
Willow==0.4
